Resize images to be the same size.
Exception: Moviepy: ImageSequenceClip requires all images to be the same size

Please make the script adjust the images to be the same size, without losing the face alignment.

You can use the cv2.resize() function to resize the images to the same size, while preserving the aspect ratio of the image, in order to avoid losing the face alignment.
You can set the size to be the size of the image with the largest width and height.
Here is an example of how the script can be modified to resize the images to the same size: