Ok more
rot_mat = cv2.getRotationMatrix2D(left_eye_center, angle, 1.0)
TypeError: Can't parse 'center'. Sequence item with index 0 has a wrong type

It seems that the error is caused by the cv2.getRotationMatrix2D() function, where it expects the center of rotation to be passed as a tuple of x and y coordinates, but it's getting a string 'center' instead.

Here's the modification you can make to the align_pupils() function to fix the error: