omg
    rot_mat = cv2.getRotationMatrix2D(left_eye_center, angle, 1.0)
TypeError: Can't parse 'center'. Sequence item with index 0 has a wrong type

The error is caused by the fact that left_eye_center is a tuple of ints, but cv2.getRotationMatrix2D() function expects it to be a tuple of floats.

Here's how you can change the code to fix the error: